refactor(xrpl): ammendment transactions on its own file
