+++ xrpl/model/transactions/amm_bid.go
@@ -1,0 +1,10 @@
+package transactions
+
+// TODO: Implement AMMBid
+type AMMBid struct {
+	BaseTx
+}
+
+func (*AMMBid) TxType() TxType {
+	return AMMBidTx
+}
