+++ xrpl/model/transactions/amm_create.go
@@ -1,0 +1,9 @@
+package transactions
+
+type AMMCreate struct {
+	BaseTx
+}
+
+func (*AMMCreate) TxType() TxType {
+	return AMMCreateTx
+}
