+++ xrpl/model/transactions/amm_deposit.go
@@ -1,0 +1,9 @@
+package transactions
+
+type AMMDeposit struct {
+	BaseTx
+}
+
+func (*AMMDeposit) TxType() TxType {
+	return AMMDepositTx
+}
