+++ xrpl/model/transactions/amm_vote.go
@@ -1,0 +1,9 @@
+package transactions
+
+type AMMVote struct {
+	BaseTx
+}
+
+func (*AMMVote) TxType() TxType {
+	return AMMVoteTx
+}
