+++ xrpl/model/transactions/amm_withdraw.go
@@ -1,0 +1,9 @@
+package transactions
+
+type AMMWithdraw struct {
+	BaseTx
+}
+
+func (*AMMWithdraw) TxType() TxType {
+	return AMMWithdrawTx
+}
